requirements.txt
SpeechRecognition
pyaudio
google-trans-new
gTTS